post-disaster-diffusion-lora/requirements.txt
# Core model and training dependencies
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0

# Install PyTorch separately for your CUDA version
# Example:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Metrics
lpips>=0.1.4
scipy>=1.11.0

# Data processing and plotting
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Optional: spreadsheet export
openpyxl>=3.1.0
